+++ scripts.js
@@ -44,6 +44,15 @@ function replaceSkin(srcFile, skin, destFile) {
     fs.writeFileSync(destFile, text, { encoding: 'utf8' });
 }
 
+function addBanner(file) {
+    if (!fs.existsSync(file)) return;
+    const contents = [
+        banner,
+        fs.readFileSync(file, { encoding: 'utf8' }),
+    ];
+    fs.writeFileSync(file, contents.join('\n\n'), { encoding: 'utf8' });
+}
+
 const scripts = {
     /** Remove any files produced by the build process */
     'clean-all': () => {
@@ -57,10 +66,18 @@ const scripts = {
         const contents = [
             banner,
             license,
-            fs.readFileSync(`build/${PACKAGE.name}-plugin.js`),
-            fs.readFileSync(`lib/${PACKAGE.name}-component-build.js`),
+            fs.readFileSync(`build/${PACKAGE.name}-plugin.js`, { encoding: 'utf8' }),
+            fs.readFileSync(`lib/${PACKAGE.name}-component-build.js`, { encoding: 'utf8' }),
         ];
         fs.writeFileSync(outputFile, contents.join('\n\n'), { encoding: 'utf8' });
+    },
+
+    /** Add a banner with version info to the built files */
+    'add-banners': () => {
+        addBanner(`build/${PACKAGE.name}-plugin.js`);
+        addBanner(`build/${PACKAGE.name}-plugin.js.LICENSE.txt`);
+        addBanner(`build/${PACKAGE.name}.css`);
+        addBanner(`build/${PACKAGE.name}-light.css`);
     },
 
     /** Prepare module files for light skin */
@@ -78,9 +95,6 @@ const scripts = {
     },
 };
 
-// scripts['clean-rubbish']();
-
-// console.log(process.argv)
 
 const parser = new argparse.ArgumentParser({ description: '' });
 parser.add_argument('script_name', { choices: Object.keys(scripts) });